Broaden wrapper coverage without monolithic matrix code
The generated matrix now covers custom wrappers and nested smart-pointer, variant, and array combinations while keeping the axis data split from generator logic. The unit test target is renamed so matrix coverage can carry integration-style combinations and the remaining unit tests can stay focused.

Constraint: The matrix generator must remain readable enough to extend as more axes are added.

Rejected: Keep all axis data in model.py | it made the generator and data definitions hard to review independently.

Rejected: Keep wrapper behavior only in dingo_unit_test | it left runtime/static/container combinations under-covered.

Confidence: medium

Scope-risk: moderate

Directive: Add new matrix coverage by extending axis data and feature checks, not by reintroducing ad hoc generator branches.

Tested: python3 -m py_compile test/matrix/*.py

Tested: git diff --cached --check

Tested: uv run --locked python test/matrix/generate.py --out /tmp/dingo-review-generated --cmake /tmp/dingo-review-generated.cmake

Tested: cmake --build /tmp/dingo-matrix-custom-build --target dingo_matrix_test -j2

Tested: /tmp/dingo-matrix-custom-build/dingo_matrix_test --gtest_brief=1

Tested: cmake --build /tmp/dingo-matrix-custom-build --target dingo_unit_test -j2

Tested: /tmp/dingo-matrix-custom-build/dingo_unit_test --gtest_brief=1

Not-tested: Full CI matrix before push
