+++ include/dingo/resolution/runtime_binding.h
@@ -108,6 +108,24 @@ class runtime_binding
     using request_type = instance_request<rtti_type>;
     using conversion_types =
         detail::runtime_binding_conversion_types_t<Type, Storage>;
+    using exact_value_types = std::conditional_t<
+        type_traits<Type>::enabled && !std::is_pointer_v<Type>,
+        type_list<Type>, type_list<>>;
+    using exact_lvalue_reference_types = std::conditional_t<
+        Storage::conversions::is_stable && type_traits<Type>::enabled &&
+            !std::is_pointer_v<Type>,
+        type_list<Type&>, type_list<>>;
+    using exact_pointer_types = std::conditional_t<
+        Storage::conversions::is_stable && type_traits<Type>::enabled &&
+            !std::is_pointer_v<Type>,
+        type_list<Type*>, type_list<>>;
+    using value_capability_types = type_list_cat_t<
+        exact_value_types, typename Storage::conversions::value_types>;
+    using lvalue_reference_capability_types = type_list_cat_t<
+        exact_lvalue_reference_types,
+        typename Storage::conversions::lvalue_reference_types>;
+    using pointer_capability_types = type_list_cat_t<
+        exact_pointer_types, typename Storage::conversions::pointer_types>;
     static constexpr bool has_conversion_cache =
         detail::runtime_binding_has_conversion_cache_v<conversion_types>;
     static constexpr bool uses_cached_conversions =
@@ -191,14 +209,13 @@ class runtime_binding
 
     void* get_value(runtime_context& context, const request_type& request,
                     instance_cache_sink cache) override {
-        return convert<typename Storage::conversions::value_types>(
-            context, request, cache);
+        return convert<value_capability_types>(context, request, cache);
     }
 
     void* get_lvalue_reference(runtime_context& context,
                                const request_type& request,
                                instance_cache_sink cache) override {
-        return convert<typename Storage::conversions::lvalue_reference_types>(
+        return convert<lvalue_reference_capability_types>(
             context, request, cache);
     }
 
@@ -211,8 +228,7 @@ class runtime_binding
 
     void* get_pointer(runtime_context& context, const request_type& request,
                       instance_cache_sink cache) override {
-        return convert<typename Storage::conversions::pointer_types>(
-            context, request, cache);
+        return convert<pointer_capability_types>(context, request, cache);
     }
 
 #ifdef _MSC_VER
