+++ test/matrix/axis_exposed_types.py
@@ -1,0 +1,424 @@
+#!/usr/bin/env python3
+#
+# This file is part of dingo project <https://github.com/romanpauk/dingo>
+#
+# See LICENSE for license and copyright information
+# SPDX-License-Identifier: MIT
+#
+
+from __future__ import annotations
+
+from schema import (
+    ContainerSpec,
+    ExposedType,
+    Feature,
+    RegistrationMode,
+    RegistrationSpec,
+    ResolvedType,
+    ScopeSpec,
+    StoredType,
+)
+
+
+EXPOSED_TYPES = (
+    ExposedType(
+        name="concrete",
+        kind="concrete",
+        supported_stored_kinds=frozenset({"value_type"}),
+        provides=frozenset({"concrete_binding"}),
+        registrations=(RegistrationSpec(),),
+    ),
+    ExposedType(
+        name="mixed_external_dependency",
+        kind="mixed_external_dependency",
+        supported_stored_kinds=frozenset({"dependent_type"}),
+        provides=frozenset({"mixed_external_dependency"}),
+        registrations=(
+            RegistrationSpec(
+                scope="dingo::scope<dingo::external>",
+                storage="dingo::storage<value_type&>",
+                runtime_setup=("value_type external_value;",),
+                runtime_argument="external_value",
+                mixed="runtime",
+                static=False,
+            ),
+            RegistrationSpec(),
+        ),
+    ),
+    ExposedType(
+        name="scenario",
+        kind="scenario",
+        supported_stored_kinds=frozenset({"scenario"}),
+        provides=frozenset({"scenario"}),
+        registrations=(),
+    ),
+    ExposedType(
+        name="interface_type",
+        kind="interface",
+        supported_stored_kinds=frozenset({"implementation_type"}),
+        provides=frozenset({"concrete_binding", "interface_binding"}),
+        registrations=(
+            RegistrationSpec(storage="dingo::storage<value_type>"),
+            RegistrationSpec(
+                interfaces="dingo::interfaces<interface_type>",
+                mixed="runtime",
+            ),
+        ),
+    ),
+    ExposedType(
+        name="multiple_interfaces",
+        kind="interface",
+        supported_stored_kinds=frozenset({"implementation_type"}),
+        provides=frozenset({"concrete_binding", "interface_binding"}),
+        registrations=(
+            RegistrationSpec(storage="dingo::storage<value_type>"),
+            RegistrationSpec(
+                interfaces="dingo::interfaces<interface_type, second_interface_type>"
+            ),
+        ),
+    ),
+    ExposedType(
+        name="copyable_interface_type",
+        kind="interface",
+        supported_stored_kinds=frozenset({"implementation_type"}),
+        provides=frozenset({"concrete_binding", "interface_binding"}),
+        registrations=(
+            RegistrationSpec(storage="dingo::storage<value_type>"),
+            RegistrationSpec(
+                interfaces="dingo::interfaces<copyable_interface_type>",
+                mixed="runtime",
+            ),
+        ),
+    ),
+    ExposedType(
+        name="custom_concrete",
+        kind="custom_concrete",
+        supported_stored_kinds=frozenset({"custom_value_type"}),
+        provides=frozenset({"custom_wrapper_binding"}),
+        registrations=(RegistrationSpec(),),
+    ),
+    ExposedType(
+        name="custom_interface",
+        kind="custom_interface",
+        supported_stored_kinds=frozenset({"custom_interface_type"}),
+        provides=frozenset({"custom_wrapper_binding"}),
+        registrations=(
+            RegistrationSpec(interfaces="dingo::interfaces<interface_type>"),
+        ),
+    ),
+    ExposedType(
+        name="custom_unique_interface",
+        kind="custom_interface",
+        supported_stored_kinds=frozenset({"custom_unique_interface_type"}),
+        provides=frozenset({"custom_wrapper_binding"}),
+        registrations=(
+            RegistrationSpec(interfaces="dingo::interfaces<unique_interface_type>"),
+        ),
+    ),
+    ExposedType(
+        name="keyed_concrete",
+        kind="keyed",
+        supported_stored_kinds=frozenset({"value_type"}),
+        provides=frozenset({"keyed_binding"}),
+        registrations=(RegistrationSpec(key="dingo::key<key_a>"),),
+    ),
+    ExposedType(
+        name="keyed_interface",
+        kind="keyed",
+        supported_stored_kinds=frozenset({"implementation_type"}),
+        provides=frozenset({"keyed_binding", "interface_binding"}),
+        registrations=(
+            RegistrationSpec(storage="dingo::storage<value_type>"),
+            RegistrationSpec(
+                interfaces="dingo::interfaces<interface_type>",
+                key="dingo::key<key_a>",
+            ),
+        ),
+    ),
+    ExposedType(
+        name="annotated_concrete",
+        kind="annotated",
+        supported_stored_kinds=frozenset({"value_type"}),
+        provides=frozenset({"annotated_binding"}),
+        registrations=(
+            RegistrationSpec(
+                interfaces="dingo::interfaces<dingo::annotated<value_type, tag_a>>"
+            ),
+        ),
+    ),
+    ExposedType(
+        name="annotated_interface",
+        kind="annotated",
+        supported_stored_kinds=frozenset({"implementation_type"}),
+        provides=frozenset({"annotated_binding", "interface_binding"}),
+        registrations=(
+            RegistrationSpec(storage="dingo::storage<value_type>"),
+            RegistrationSpec(
+                interfaces="dingo::interfaces<dingo::annotated<interface_type, tag_a>>",
+            ),
+        ),
+    ),
+    ExposedType(
+        name="array_concrete",
+        kind="array",
+        supported_stored_kinds=frozenset({"array_type"}),
+        provides=frozenset({"array_binding"}),
+        registrations=(RegistrationSpec(),),
+    ),
+    ExposedType(
+        name="variant_concrete",
+        kind="variant",
+        supported_stored_kinds=frozenset({"variant_type"}),
+        provides=frozenset({"variant_binding"}),
+        registrations=(
+            RegistrationSpec(storage="dingo::storage<value_type>"),
+            RegistrationSpec(),
+        ),
+    ),
+    ExposedType(
+        name="nested_wrapper",
+        kind="nested_wrapper",
+        supported_stored_kinds=frozenset({"nested_wrapper_type"}),
+        provides=frozenset({"nested_wrapper_binding"}),
+        registrations=(RegistrationSpec(),),
+    ),
+    ExposedType(
+        name="variant_unique_wrapper",
+        kind="nested_wrapper",
+        supported_stored_kinds=frozenset({"variant_unique_wrapper_type"}),
+        provides=frozenset({"nested_wrapper_binding"}),
+        registrations=(RegistrationSpec(),),
+    ),
+    ExposedType(
+        name="variant_shared_wrapper",
+        kind="nested_wrapper",
+        supported_stored_kinds=frozenset({"variant_shared_wrapper_type"}),
+        provides=frozenset({"nested_wrapper_binding"}),
+        registrations=(RegistrationSpec(),),
+    ),
+    ExposedType(
+        name="variant_shared_unique_wrapper",
+        kind="nested_wrapper",
+        supported_stored_kinds=frozenset({"variant_shared_unique_wrapper_type"}),
+        provides=frozenset({"nested_wrapper_binding"}),
+        registrations=(RegistrationSpec(),),
+    ),
+    ExposedType(
+        name="local_value",
+        kind="local_value_type",
+        supported_stored_kinds=frozenset({"local_value_type"}),
+        provides=frozenset({"local_bindings"}),
+        registrations=(
+            RegistrationSpec(storage="dingo::storage<value_type>"),
+            RegistrationSpec(
+                local_bindings=(
+                    "dingo::bindings<dingo::bind<dingo::scope<dingo::shared>, "
+                    "dingo::storage<local_dependency_type>>>"
+                )
+            ),
+        ),
+    ),
+    ExposedType(
+        name="local_override_value",
+        kind="local_override_value_type",
+        supported_stored_kinds=frozenset({"local_override_value_type"}),
+        provides=frozenset({"local_bindings"}),
+        registrations=(
+            RegistrationSpec(
+                storage="dingo::storage<local_dependency_type>",
+                factory="dingo::factory<dingo::constructor<local_dependency_type()>>",
+            ),
+            RegistrationSpec(
+                local_bindings=(
+                    "dingo::bindings<dingo::bind<dingo::scope<dingo::shared>, "
+                    "dingo::storage<local_dependency_type>>>"
+                )
+            ),
+        ),
+    ),
+    ExposedType(
+        name="local_collection_value",
+        kind="local_collection_value_type",
+        supported_stored_kinds=frozenset({"local_collection_value_type"}),
+        provides=frozenset({"local_bindings"}),
+        registrations=(
+            RegistrationSpec(
+                storage="dingo::storage<std::shared_ptr<element_type<1>>>",
+                interfaces="dingo::interfaces<element_interface>",
+            ),
+            RegistrationSpec(
+                dependencies="dingo::dependencies<std::vector<std::shared_ptr<element_interface>>>",
+                local_bindings=(
+                    "dingo::bindings<dingo::bind<dingo::scope<dingo::shared>, "
+                    "dingo::storage<std::shared_ptr<element_type<0>>>, "
+                    "dingo::interfaces<element_interface>>>"
+                )
+            ),
+        ),
+    ),
+    ExposedType(
+        name="local_collection_runtime_host_value",
+        kind="local_collection_value_type",
+        supported_stored_kinds=frozenset({"local_collection_value_type"}),
+        provides=frozenset({"local_bindings"}),
+        registrations=(
+            RegistrationSpec(
+                storage="dingo::storage<std::shared_ptr<element_type<1>>>",
+                interfaces="dingo::interfaces<element_interface>",
+                mixed="runtime",
+            ),
+            RegistrationSpec(
+                dependencies="dingo::dependencies<std::vector<std::shared_ptr<element_interface>>>",
+                local_bindings=(
+                    "dingo::bindings<dingo::bind<dingo::scope<dingo::shared>, "
+                    "dingo::storage<std::shared_ptr<element_type<0>>>, "
+                    "dingo::interfaces<element_interface>>>"
+                ),
+            ),
+        ),
+    ),
+    ExposedType(
+        name="factory_function_value",
+        kind="factory_function_value_type",
+        supported_stored_kinds=frozenset({"factory_function_value_type"}),
+        provides=frozenset({"factory_override"}),
+        registrations=(
+            RegistrationSpec(
+                factory="dingo::factory<dingo::function<&factory_function_value_type::create>>"
+            ),
+        ),
+    ),
+    ExposedType(
+        name="factory_constructor_value",
+        kind="factory_constructor_value_type",
+        supported_stored_kinds=frozenset({"factory_constructor_value_type"}),
+        provides=frozenset({"factory_override"}),
+        registrations=(
+            RegistrationSpec(storage="dingo::storage<value_type>"),
+            RegistrationSpec(
+                factory="dingo::factory<dingo::constructor<factory_constructor_value_type(value_type&)>>"
+            ),
+        ),
+    ),
+    ExposedType(
+        name="cycle_concrete",
+        kind="cycle",
+        supported_stored_kinds=frozenset({"cycle_type"}),
+        provides=frozenset({"concrete_binding"}),
+        registrations=(
+            RegistrationSpec(),
+            RegistrationSpec(storage="dingo::storage<std::shared_ptr<cycle_b_type>>"),
+        ),
+    ),
+    ExposedType(
+        name="unique_interface_type",
+        kind="interface",
+        supported_stored_kinds=frozenset({"unique_implementation_type"}),
+        provides=frozenset({"interface_binding"}),
+        registrations=(
+            RegistrationSpec(interfaces="dingo::interfaces<unique_interface_type>"),
+        ),
+    ),
+    ExposedType(
+        name="element_collection",
+        kind="collection",
+        supported_stored_kinds=frozenset({"element_type"}),
+        provides=frozenset({"collection_binding"}),
+        registrations=(
+            RegistrationSpec(
+                storage="dingo::storage<std::shared_ptr<element_type<0>>>",
+                interfaces="dingo::interfaces<element_interface>",
+            ),
+            RegistrationSpec(
+                storage="dingo::storage<std::shared_ptr<element_type<1>>>",
+                interfaces="dingo::interfaces<element_interface>",
+                mixed="runtime",
+            ),
+        ),
+        runtime_prefix=(
+            "container.template register_type_collection<"
+            "dingo::scope<dingo::unique>, "
+            "dingo::storage<std::vector<std::shared_ptr<"
+            "element_interface>>>>();",
+        ),
+        mixed_runtime_prefix=(
+            "container.template register_type_collection<"
+            "dingo::scope<dingo::unique>, "
+            "dingo::storage<std::vector<std::shared_ptr<"
+            "element_interface>>>>();",
+        ),
+    ),
+    ExposedType(
+        name="element_keyed",
+        kind="keyed",
+        supported_stored_kinds=frozenset({"element_type"}),
+        provides=frozenset({"keyed_binding"}),
+        registrations=(
+            RegistrationSpec(
+                storage="dingo::storage<std::shared_ptr<element_type<0>>>",
+                interfaces="dingo::interfaces<element_interface>",
+                key="dingo::key<key_a>",
+            ),
+            RegistrationSpec(
+                storage="dingo::storage<std::shared_ptr<element_type<1>>>",
+                interfaces="dingo::interfaces<element_interface>",
+                key="dingo::key<key_b>",
+                mixed="runtime",
+            ),
+        ),
+    ),
+    ExposedType(
+        name="element_keyed_collection",
+        kind="keyed_collection",
+        supported_stored_kinds=frozenset({"element_type"}),
+        provides=frozenset({"keyed_collection_binding"}),
+        registrations=(
+            RegistrationSpec(
+                storage="dingo::storage<std::shared_ptr<element_type<0>>>",
+                interfaces="dingo::interfaces<element_interface>",
+                key="dingo::key<key_a>",
+            ),
+            RegistrationSpec(
+                storage="dingo::storage<std::shared_ptr<element_type<1>>>",
+                interfaces="dingo::interfaces<element_interface>",
+                key="dingo::key<key_a>",
+                mixed="runtime",
+            ),
+        ),
+        runtime_prefix=(
+            "container.template register_type_collection<"
+            "dingo::scope<dingo::unique>, "
+            "dingo::storage<std::vector<std::shared_ptr<"
+            "element_interface>>>>(dingo::key<key_a>{});",
+        ),
+        mixed_runtime_prefix=(
+            "container.template register_type_collection<"
+            "dingo::scope<dingo::unique>, "
+            "dingo::storage<std::vector<std::shared_ptr<"
+            "element_interface>>>>(dingo::key<key_a>{});",
+        ),
+    ),
+    ExposedType(
+        name="element_indexed",
+        kind="indexed",
+        supported_stored_kinds=frozenset({"element_type"}),
+        provides=frozenset({"indexed_binding"}),
+        registrations=(
+            RegistrationSpec(
+                storage="dingo::storage<std::shared_ptr<element_type<0>>>",
+                interfaces="dingo::interfaces<element_interface>",
+                indexed_key="std::size_t{0}",
+                mixed="runtime",
+                static=False,
+            ),
+            RegistrationSpec(
+                storage="dingo::storage<std::shared_ptr<element_type<1>>>",
+                interfaces="dingo::interfaces<element_interface>",
+                indexed_key="std::size_t{1}",
+                mixed="runtime",
+                static=False,
+            ),
+        ),
+        mixed=False,
+    ),
+)
