+++ test/matrix/axis_stored_types.py
@@ -1,0 +1,439 @@
+#!/usr/bin/env python3
+#
+# This file is part of dingo project <https://github.com/romanpauk/dingo>
+#
+# See LICENSE for license and copyright information
+# SPDX-License-Identifier: MIT
+#
+
+from __future__ import annotations
+
+from schema import (
+    ContainerSpec,
+    ExposedType,
+    Feature,
+    RegistrationMode,
+    RegistrationSpec,
+    ResolvedType,
+    ScopeSpec,
+    StoredType,
+)
+
+
+STORED_TYPES = (
+    StoredType(
+        id="value_type",
+        name="value_type",
+        kind="value_type",
+        storage="dingo::storage<value_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset(
+            {"stored_value", "direct_value_resolution", "lifetime_countable"}
+        ),
+    ),
+    StoredType(
+        id="unique_value_type",
+        name="value_type",
+        kind="value_type",
+        storage="dingo::storage<value_type>",
+        supported_scopes=frozenset({"unique"}),
+        provides=frozenset(
+            {"stored_value", "direct_value_resolution", "lifetime_countable"}
+        ),
+    ),
+    StoredType(
+        id="dependent_type",
+        name="dependent_type",
+        kind="dependent_type",
+        storage="dingo::storage<dependent_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_value"}),
+    ),
+    StoredType(
+        id="scenario",
+        name="scenario_type",
+        kind="scenario",
+        storage="dingo::storage<scenario_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"scenario"}),
+    ),
+    StoredType(
+        id="external_value_type",
+        name="external<value_type>",
+        kind="value_type",
+        storage="dingo::storage<value_type>",
+        supported_scopes=frozenset({"external"}),
+        provides=frozenset({"stored_external_value", "direct_value_resolution"}),
+        supported_modes=frozenset({"runtime"}),
+        runtime_setup=("value_type external_value;",),
+        runtime_argument="std::move(external_value)",
+    ),
+    StoredType(
+        id="external_value_ref",
+        name="external<value_type&>",
+        kind="value_type",
+        storage="dingo::storage<value_type&>",
+        supported_scopes=frozenset({"external"}),
+        provides=frozenset({"stored_external_reference", "direct_value_resolution"}),
+        supported_modes=frozenset({"runtime"}),
+        runtime_setup=("value_type external_value;",),
+        runtime_argument="external_value",
+    ),
+    StoredType(
+        id="external_value_ptr",
+        name="external<value_type*>",
+        kind="value_type",
+        storage="dingo::storage<value_type*>",
+        supported_scopes=frozenset({"external"}),
+        provides=frozenset({"stored_external_pointer", "direct_value_resolution"}),
+        supported_modes=frozenset({"runtime"}),
+        runtime_setup=("value_type external_value;",),
+        runtime_argument="&external_value",
+    ),
+    StoredType(
+        id="value_ptr",
+        name="value_type*",
+        kind="value_type",
+        storage="dingo::storage<value_type*>",
+        supported_scopes=frozenset({"external"}),
+        provides=frozenset({"stored_pointer", "direct_value_resolution"}),
+        supported_modes=frozenset({"runtime"}),
+        runtime_setup=("value_type external_value;",),
+        runtime_argument="&external_value",
+    ),
+    StoredType(
+        id="value_ref",
+        name="value_type&",
+        kind="value_type",
+        storage="dingo::storage<value_type&>",
+        supported_scopes=frozenset({"external"}),
+        provides=frozenset({"stored_reference", "direct_value_resolution"}),
+        supported_modes=frozenset({"runtime"}),
+        runtime_setup=("value_type external_value;",),
+        runtime_argument="external_value",
+    ),
+    StoredType(
+        id="external_value_shared_ptr",
+        name="external<std::shared_ptr<value_type>>",
+        kind="value_type",
+        storage="dingo::storage<std::shared_ptr<value_type>>",
+        supported_scopes=frozenset({"external"}),
+        provides=frozenset({"stored_external_shared_ptr", "direct_value_resolution"}),
+        supported_modes=frozenset({"runtime"}),
+        runtime_setup=("auto external_value = std::make_shared<value_type>();",),
+        runtime_argument="external_value",
+    ),
+    StoredType(
+        id="value_unique_ptr",
+        name="std::unique_ptr<value_type>",
+        kind="value_type",
+        storage="dingo::storage<std::unique_ptr<value_type>>",
+        supported_scopes=frozenset({"unique"}),
+        provides=frozenset({"stored_unique_ptr", "lifetime_countable"}),
+    ),
+    StoredType(
+        id="value_shared_ptr",
+        name="std::shared_ptr<value_type>",
+        kind="value_type",
+        storage="dingo::storage<std::shared_ptr<value_type>>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_shared_ptr", "lifetime_countable"}),
+    ),
+    StoredType(
+        id="value_optional",
+        name="std::optional<value_type>",
+        kind="value_type",
+        storage="dingo::storage<std::optional<value_type>>",
+        supported_scopes=frozenset({"shared", "unique"}),
+        provides=frozenset({"stored_optional", "lifetime_countable"}),
+    ),
+    StoredType(
+        id="custom_shared_value",
+        name="dingo::test_shared<value_type>",
+        kind="custom_value_type",
+        storage="dingo::storage<dingo::test_shared<value_type>>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_custom_shared"}),
+    ),
+    StoredType(
+        id="external_custom_shared_value",
+        name="external<dingo::test_shared<value_type>>",
+        kind="custom_value_type",
+        storage="dingo::storage<dingo::test_shared<value_type>>",
+        supported_scopes=frozenset({"external"}),
+        provides=frozenset({"stored_custom_shared"}),
+        supported_modes=frozenset({"runtime"}),
+        runtime_setup=(
+            "auto external_value = dingo::type_traits<dingo::test_shared<value_type>>::make();",
+        ),
+        runtime_argument="std::move(external_value)",
+    ),
+    StoredType(
+        id="external_custom_shared_value_ref",
+        name="external<dingo::test_shared<value_type>&>",
+        kind="custom_value_type",
+        storage="dingo::storage<dingo::test_shared<value_type>&>",
+        supported_scopes=frozenset({"external"}),
+        provides=frozenset({"stored_custom_shared"}),
+        supported_modes=frozenset({"runtime"}),
+        runtime_setup=(
+            "auto external_value = dingo::type_traits<dingo::test_shared<value_type>>::make();",
+        ),
+        runtime_argument="external_value",
+    ),
+    StoredType(
+        id="custom_shared_interface",
+        name="dingo::test_shared<custom_interface_implementation_type>",
+        kind="custom_interface_type",
+        storage=(
+            "dingo::storage<dingo::test_shared<"
+            "custom_interface_implementation_type>>"
+        ),
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_custom_shared"}),
+    ),
+    StoredType(
+        id="custom_unique_interface",
+        name="dingo::test_unique<custom_unique_implementation_type>",
+        kind="custom_unique_interface_type",
+        storage=(
+            "dingo::storage<dingo::test_unique<"
+            "custom_unique_implementation_type>>"
+        ),
+        supported_scopes=frozenset({"unique"}),
+        provides=frozenset({"stored_custom_unique"}),
+    ),
+    StoredType(
+        id="custom_optional_value",
+        name="dingo::test_optional<value_type>",
+        kind="custom_value_type",
+        storage="dingo::storage<dingo::test_optional<value_type>>",
+        supported_scopes=frozenset({"shared", "unique"}),
+        provides=frozenset({"stored_custom_optional"}),
+    ),
+    StoredType(
+        id="external_custom_optional_value",
+        name="external<dingo::test_optional<value_type>>",
+        kind="custom_value_type",
+        storage="dingo::storage<dingo::test_optional<value_type>>",
+        supported_scopes=frozenset({"external"}),
+        provides=frozenset({"stored_custom_optional"}),
+        supported_modes=frozenset({"runtime"}),
+        runtime_setup=(
+            "auto external_value = dingo::type_traits<dingo::test_optional<value_type>>::make();",
+        ),
+        runtime_argument="std::move(external_value)",
+    ),
+    StoredType(
+        id="external_custom_optional_value_ref",
+        name="external<dingo::test_optional<value_type>&>",
+        kind="custom_value_type",
+        storage="dingo::storage<dingo::test_optional<value_type>&>",
+        supported_scopes=frozenset({"external"}),
+        provides=frozenset({"stored_custom_optional"}),
+        supported_modes=frozenset({"runtime"}),
+        runtime_setup=(
+            "auto external_value = dingo::type_traits<dingo::test_optional<value_type>>::make();",
+        ),
+        runtime_argument="external_value",
+    ),
+    StoredType(
+        id="value_array_2",
+        name="value_type[2]",
+        kind="array_type",
+        storage="dingo::storage<value_type[2]>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_raw_array", "array_1d"}),
+        supported_modes=frozenset({"runtime", "static", "mixed"}),
+    ),
+    StoredType(
+        id="value_array_2_3",
+        name="value_type[2][3]",
+        kind="array_type",
+        storage="dingo::storage<value_type[2][3]>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_raw_array", "array_2d"}),
+        supported_modes=frozenset({"runtime", "static", "mixed"}),
+    ),
+    StoredType(
+        id="value_unique_array",
+        name="std::unique_ptr<value_type[]>",
+        kind="array_type",
+        storage="dingo::storage<std::unique_ptr<value_type[]>>",
+        supported_scopes=frozenset({"unique"}),
+        provides=frozenset({"stored_unique_array", "array_smart_unique"}),
+        factory="dingo::factory<dingo::function<&make_unique_value_array>>",
+    ),
+    StoredType(
+        id="value_shared_array",
+        name="std::shared_ptr<value_type[]>",
+        kind="array_type",
+        storage="dingo::storage<std::shared_ptr<value_type[]>>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_shared_array", "array_smart_shared"}),
+        factory="dingo::factory<dingo::function<&make_shared_value_array>>",
+    ),
+    StoredType(
+        id="value_variant",
+        name="std::variant<variant_a, variant_b>",
+        kind="variant_type",
+        storage="dingo::storage<std::variant<variant_a, variant_b>>",
+        supported_scopes=frozenset({"unique", "shared"}),
+        provides=frozenset({"stored_variant"}),
+        factory="dingo::factory<dingo::constructor<variant_a(value_type&)>>",
+    ),
+    StoredType(
+        id="shared_unique_value",
+        name="std::shared_ptr<std::unique_ptr<value_type>>",
+        kind="nested_wrapper_type",
+        storage="dingo::storage<shared_unique_value_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_shared_unique"}),
+        factory="dingo::factory<dingo::function<&make_shared_unique_value>>",
+    ),
+    StoredType(
+        id="shared_unique_array_value",
+        name="std::shared_ptr<std::unique_ptr<value_type[]>>",
+        kind="nested_wrapper_type",
+        storage="dingo::storage<shared_unique_array_value_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_shared_unique_array"}),
+        factory="dingo::factory<dingo::function<&make_shared_unique_array_value>>",
+    ),
+    StoredType(
+        id="unique_shared_value",
+        name="std::unique_ptr<std::shared_ptr<value_type>>",
+        kind="nested_wrapper_type",
+        storage="dingo::storage<unique_shared_value_type>",
+        supported_scopes=frozenset({"unique"}),
+        provides=frozenset({"stored_unique_shared"}),
+        factory="dingo::factory<dingo::function<&make_unique_shared_value>>",
+    ),
+    StoredType(
+        id="variant_unique_value",
+        name="std::variant<std::unique_ptr<value_type>, nested_variant_b>",
+        kind="variant_unique_wrapper_type",
+        storage="dingo::storage<variant_unique_value_type>",
+        supported_scopes=frozenset({"unique"}),
+        provides=frozenset({"stored_variant_unique"}),
+        factory="dingo::factory<dingo::function<&make_variant_unique_value>>",
+    ),
+    StoredType(
+        id="variant_shared_value",
+        name="std::variant<std::shared_ptr<value_type>, nested_variant_b>",
+        kind="variant_shared_wrapper_type",
+        storage="dingo::storage<variant_shared_value_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_variant_shared"}),
+        factory="dingo::factory<dingo::function<&make_variant_shared_value>>",
+    ),
+    StoredType(
+        id="variant_shared_unique_value",
+        name="std::variant<std::shared_ptr<std::unique_ptr<value_type>>, nested_variant_b>",
+        kind="variant_shared_unique_wrapper_type",
+        storage="dingo::storage<variant_shared_unique_value_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_variant_shared_unique"}),
+        factory="dingo::factory<dingo::function<&make_variant_shared_unique_value>>",
+    ),
+    StoredType(
+        id="unique_variant_value",
+        name="std::unique_ptr<std::variant<nested_variant_a, nested_variant_b>>",
+        kind="nested_wrapper_type",
+        storage="dingo::storage<unique_variant_value_type>",
+        supported_scopes=frozenset({"unique"}),
+        provides=frozenset({"stored_unique_variant"}),
+        factory="dingo::factory<dingo::function<&make_unique_variant_value>>",
+    ),
+    StoredType(
+        id="shared_variant_value",
+        name="std::shared_ptr<std::variant<nested_variant_a, nested_variant_b>>",
+        kind="nested_wrapper_type",
+        storage="dingo::storage<shared_variant_value_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_shared_variant"}),
+        factory="dingo::factory<dingo::function<&make_shared_variant_value>>",
+    ),
+    StoredType(
+        id="array_variant_value",
+        name="std::variant<std::array<value_type, 2>, nested_variant_b>",
+        kind="nested_wrapper_type",
+        storage="dingo::storage<array_variant_value_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_array_variant"}),
+        factory="dingo::factory<dingo::function<&make_array_variant_value>>",
+    ),
+    StoredType(
+        id="local_value_type",
+        name="local_value_type",
+        kind="local_value_type",
+        storage="dingo::storage<local_value_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_value"}),
+    ),
+    StoredType(
+        id="local_override_value_type",
+        name="local_override_value_type",
+        kind="local_override_value_type",
+        storage="dingo::storage<local_override_value_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_value"}),
+    ),
+    StoredType(
+        id="local_collection_value_type",
+        name="local_collection_value_type",
+        kind="local_collection_value_type",
+        storage="dingo::storage<local_collection_value_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_value"}),
+    ),
+    StoredType(
+        id="factory_function_value_type",
+        name="factory_function_value_type",
+        kind="factory_function_value_type",
+        storage="dingo::storage<factory_function_value_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_value"}),
+    ),
+    StoredType(
+        id="factory_constructor_value_type",
+        name="factory_constructor_value_type",
+        kind="factory_constructor_value_type",
+        storage="dingo::storage<factory_constructor_value_type>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_value"}),
+    ),
+    StoredType(
+        id="cycle_a_type",
+        name="cycle_a_type",
+        kind="cycle_type",
+        storage="dingo::storage<cycle_a_type>",
+        supported_scopes=frozenset({"shared_cyclical"}),
+        provides=frozenset({"stored_value", "cycle_graph"}),
+        supported_modes=frozenset({"runtime"}),
+    ),
+    StoredType(
+        id="implementation_shared_ptr",
+        name="std::shared_ptr<implementation_type>",
+        kind="implementation_type",
+        storage="dingo::storage<std::shared_ptr<implementation_type>>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_shared_ptr"}),
+    ),
+    StoredType(
+        id="element_shared_ptr",
+        name="std::shared_ptr<element_type<0>>",
+        kind="element_type",
+        storage="dingo::storage<std::shared_ptr<element_type<0>>>",
+        supported_scopes=frozenset({"shared"}),
+        provides=frozenset({"stored_shared_ptr"}),
+    ),
+    StoredType(
+        id="unique_implementation_unique_ptr",
+        name="std::unique_ptr<unique_implementation_type>",
+        kind="unique_implementation_type",
+        storage="dingo::storage<std::unique_ptr<unique_implementation_type>>",
+        supported_scopes=frozenset({"unique"}),
+        provides=frozenset({"stored_unique_ptr"}),
+    ),
+)
