+++ test/matrix/data.py
@@ -1,0 +1,28 @@
+#!/usr/bin/env python3
+#
+# This file is part of dingo project <https://github.com/romanpauk/dingo>
+#
+# See LICENSE for license and copyright information
+# SPDX-License-Identifier: MIT
+#
+
+from __future__ import annotations
+
+from axis_config import FILTER_RULES, REGISTRATION_MODES, SCOPES
+from axis_containers import CONTAINERS
+from axis_exposed_types import EXPOSED_TYPES
+from axis_features import FEATURES
+from axis_resolved_types import RESOLVED_TYPES
+from axis_stored_types import STORED_TYPES
+
+
+__all__ = (
+    "CONTAINERS",
+    "EXPOSED_TYPES",
+    "FEATURES",
+    "FILTER_RULES",
+    "REGISTRATION_MODES",
+    "RESOLVED_TYPES",
+    "SCOPES",
+    "STORED_TYPES",
+)
